Project-6/backend/requirements.txt
flask==2.3.0
langchain==0.1.0
google-generativeai==0.3.0
python-dotenv==1.0.0
flask-cors==4.0.0
pillow==10.0.0
requests==2.31.0
